+++ core/src/options/from_derive.rs
@@ -4,7 +4,9 @@ use syn::Ident;
 
 use crate::codegen::FromDeriveInputImpl;
 use crate::options::{DeriveInputShapeSet, OuterFrom, ParseAttribute, ParseData};
-use crate::{FromMeta, Result};
+use crate::{FromField, FromMeta, Result};
+
+use super::forwarded_field::ForwardedField;
 
 #[derive(Debug)]
 pub struct FdiOptions {
@@ -16,7 +18,8 @@ pub struct FdiOptions {
     /// The field on the target struct which should receive the type generics, if any.
     pub generics: Option<Ident>,
 
-    pub data: Option<Ident>,
+    /// The field on the target struct which should receive the derive input body, if any.
+    pub data: Option<ForwardedField>,
 
     pub supports: Option<DeriveInputShapeSet>,
 }
@@ -58,7 +61,7 @@ impl ParseData for FdiOptions {
                 Ok(())
             }
             Some("data") => {
-                self.data.clone_from(&field.ident);
+                self.data = ForwardedField::from_field(field).map(Some)?;
                 Ok(())
             }
             Some("generics") => {