+++ compiletests.sh
@@ -1,1 +1,1 @@
-RUSTFLAGS="--cfg=compiletests" cargo +1.65.0 test --test compiletests
+RUSTFLAGS="--cfg=compiletests" cargo +1.77.0 test --test compiletests